Added input processing
The game receives a key and character string from the main loop
It currently passes the key parameter to the current scene.

Scenes look at their action map to find a handler for the key.
Handlers are defined by windows in their getActions() method.

This method returns a list of key/action tuples. The scene
throws an exception if multiple windows try to listen for
the same keys.